feat: :sparkles: Pulled configs out of code, setup proper logging, refactored the python code.
Configurations are now all in the conf folder and get read in where needed. They are approximately subdivided into the task they are needed for and are integrated in the Makefile. Changing the conf for doing the analysis for example will not restart the whole build. Python logging is integrated and allow for console and file logging. Logs are saved in output/logs folder.  Finally, the python scripts are refactored so the main thing that happens is at the top of the script for better readability.

None